+++ src/youtube_automation/domains/uploads/_complete_collection_executor.py
@@ -14,9 +14,9 @@ import logging
 from pathlib import Path
 
 from youtube_automation.core.adapters import cost_tracker
-from youtube_automation.core.adapters.errors import AutomationError, QuotaExhaustedError
 from youtube_automation.core.adapters.security import redact_sensitive_data
 from youtube_automation.core.adapters.youtube import complete_collection_quota_plan, quota_shortages
+from youtube_automation.core.errors import AutomationError, QuotaExhaustedError
 from youtube_automation.domains.uploads._collection_uploader_constants import (
     ACTION_COMPLETE_COLLECTION_DEDUP_SKIPPED,
     ACTION_COMPLETE_COLLECTION_QUOTA_EXHAUSTED,
