+++ src/youtube_automation/domains/uploads/youtube.py
@@ -7,7 +7,8 @@ from pathlib import Path
 from typing import Callable, Dict, List, Optional
 
 from youtube_automation.configuration import channel_dir, load_config
-from youtube_automation.core.adapters.errors import (
+from youtube_automation.core.adapters.runtime import resolve_default_publish_at as _resolve_default_publish_at
+from youtube_automation.core.errors import (
     AutomationError,
     ConfigError,
     QuotaExhaustedError,
@@ -15,15 +16,6 @@ from youtube_automation.core.adapters.errors import (
     ValidationError,
     YouTubeAPIError,
 )
-from youtube_automation.core.adapters.filesystem import file_size, path_exists, remove_file
-from youtube_automation.core.adapters.google.upload import HttpError, create_media_upload
-from youtube_automation.core.adapters.google.youtube import (
-    YouTubeClients,
-    execute_youtube_request,
-    validate_youtube_response_items,
-)
-from youtube_automation.core.adapters.process import compress_image
-from youtube_automation.core.adapters.runtime import resolve_default_publish_at as _resolve_default_publish_at
 from youtube_automation.domains.metadata import BAHMetadataGenerator
 from youtube_automation.domains.uploads._complete_collection_strategy import CompleteCollectionMixin
 from youtube_automation.domains.uploads._dedup_search import DedupSearchMixin
@@ -37,6 +29,14 @@ from youtube_automation.domains.uploads._uploader_constants import (
 from youtube_automation.domains.uploads.policy import SESSION_EXPIRED_HTTP_STATUSES, RetryDecision, ThumbnailCompression
 from youtube_automation.domains.uploads.preflight import check_title_codepoint_limit
 from youtube_automation.domains.youtube.channel_settings import build_upload_status_flags
+from youtube_automation.infrastructure.filesystem import file_size, path_exists, remove_file
+from youtube_automation.infrastructure.google.upload import HttpError, create_media_upload
+from youtube_automation.infrastructure.google.youtube import (
+    YouTubeClients,
+    execute_youtube_request,
+    validate_youtube_response_items,
+)
+from youtube_automation.infrastructure.process import compress_image
 
 logger = logging.getLogger(__name__)
 
@@ -384,7 +384,7 @@ class YouTubeAutoUploader(
             Dict: アップロード結果
         """
         collection_dir = Path(collection_path)
-        from youtube_automation.core.adapters.filesystem import path_exists
+        from youtube_automation.infrastructure.filesystem import path_exists
 
         if not path_exists(collection_dir):
             raise FileNotFoundError(f"コレクションディレクトリが見つかりません: {collection_path}")
@@ -490,7 +490,7 @@ class YouTubeAutoUploader(
 
         for status in status_filter:
             status_dir = self.collections_root / status
-            from youtube_automation.core.adapters.filesystem import list_directory, path_exists, path_is_directory
+            from youtube_automation.infrastructure.filesystem import list_directory, path_exists, path_is_directory
 
             if path_exists(status_dir):
                 collections = [
